Test int16

+++ tests/test_minmax.py
@@ -151,7 +151,7 @@ def test_minimax_1d_negative_stride(dtype):
     assert max_val == np.amax(arr)
 
 
-@pytest.mark.parametrize("dtype", [np.float32])
+@pytest.mark.parametrize("dtype", [np.int16])
 def test_minimax_1d_non_contiguous_negative_stride(dtype):
     np.random.seed(7)
     arr = np.random.uniform(low=-6.0, high=3.0, size=(61,)).astype(dtype)[::-2]
